hookify
